Create `subspace_service::config` module
