+++ crates/subspace-node/src/commands/run.rs
@@ -36,7 +36,8 @@ use std::path::PathBuf;
 use std::sync::Arc;
 use subspace_networking::libp2p::Multiaddr;
 use subspace_runtime::{Block, ExecutorDispatch, RuntimeApi};
-use subspace_service::{DsnConfig, SubspaceConfiguration, SubspaceNetworking};
+use subspace_service::config::{SubspaceConfiguration, SubspaceNetworking};
+use subspace_service::dsn::DsnConfig;
 use tokio::runtime::Handle;
 use tracing::{debug, error, info, info_span, warn, Span};
 use tracing_subscriber::filter::LevelFilter;
